Add support for crypto unit in picocoremx7ulp
The crypto unit always occurs a kernel panic if it´s enabled. In
consultation with nxp there is a small bug. We have to enable
CONFIG_SOC_IMX6Q to support crypto unit. This is fixed in kernel
4.14.98. There we can only add CONFIG_SOC_IMX7ULP and the crypto unit
will initialize. Until then we add the CONFIG_SOC_IMX6Q to our
defconfig.