Disable user space firmware loader helper from F&S defconfigs
The user space firmware loader helper was removed for quite some time.
Remove support for it from the kernel configuration as this brings
unnecessary timeouts for devices that need a firmware, for example
for Atmel MXT touches.

While we are at it, the file permissions for atmel,maxtouch.txt in
the Documentation subtree should be 644, not 755. Fix this.