Add devicetree for GAR2
GAR2 is a customer specific board based on i.MX6UL. Add a device tree
for it. GAR2 is similar to GAR1, but has an additional Marvell 88E6071
port switch. The port is initialized in U-Boot and does not need any
attention in Linux. But of course the FEC configuration in the device
tree differs slightly.