Add RTC IRQ settings on efusA9 and efusA9X
On m ost F&S boards, the Real Time Clock (RTC) has an interrupt output
that can be used as an interrupt to the CPU, for example to wake up the
CPU or to signal a timer alarm. On efusA9 and efusA9X, the configurations
in the device trees were missing.