Improve PicoCoreMX6SX/UL device-tree - VLCD_ON pin
The VLCD_ON pin is by default active high. So setup an internal
pull-down because it will be enabled by regulator.
Note:
There is a HW bug on PicoCoreBBRGB revision 1.20 is a HW bug, VLCD_ON
provides the voltage for pca9632. So if the pca9632 is initializing,
VLCD_ON is low and the initialization fails. There is a hotfix to
prevent this. The VLCD_ON pin must be internally configured with a
pull-up so we have a high at startup and the chip can be initialized.