Add support for M12 touch variants to edt-ft5x06.c
A customer has posted a variant for the touch driver for EDT displays
edt-ft5x06.c that also supports a newer version of the touch controller,
which is called M12 by EDT. The controller is actually a Hycon 46xx. But
the protocol seems to be rather similar to the FocalTech controllers, so
that some rather small enhancements to the existing FocalTech driver
are sufficient.

The customer's version was posted here.

https://forum.fs-net.de/index.php?thread/4277-edt-touch-driver/&postID=14450#post14450

A rather similar version was added to Linux kernel 4.15 and slightly modifed
in 4.16. In fact our version here is nearly the version from 4.16, just
devm_device_add_group() can not be used in kernel 4.9 yet and needs to
stay at sysfs_create_group().