Improve USB settings of picocoremx7ulp.dts
There is a problem if we are using the dedicated USB ID pin mux option.
If we are suspend-to-RAM and resume the board, the internal USB
periphery recognize it as Host if nothing is connected and we get an
error while it trys to request the pin. There must be something wrong
with the internal USB perpiphery.
In NXP device-tree they are using an extcon-usb-gpio driver which
creates a virtual device to handle the different USB ID pin states. If
we are using this instead of dedicated id pin then everything works
fine.