Add LVDS clock source and audio clock speed selection to fsimx6 boards
Allow setting the LVDS clock source in the device tree by choosing a
predefined config macro in the dts file. There are three options:

1. PLL2_PFD0

This is a good choice for displays with pixel clock rates between 38.7 MHz
and 113 MHz. Then PLL5 can be left off, which saves energy or PLL5 can be
used by a different display (HDMI, LCD, other LVDS channel).

2. PLL5 (Video PLL)

Use the Video PLL. This PLL can be programmed to any clock rate. However
other displays (HDMI, LCD) may also want to use this PLL.

3. Use U-Boot settings

By not defining the macro at all, the settings from U-Boot are used.
Please note that the reset value is the mmdc_ch1 memory clock, so make
sure to only use this setting if U-Boot actually sets a meaningful clock
source. Otherwise the memory clock speed is changed when configuring the
LVDS display in Linux, resulting in poor overall board performance.

Example:

On efusA9, use two displays with different resolutions. Use PLL5 for
LVDS0 and PLL2_PFD0 for LVDS1.

 //#define CONFIG_EFUSA9_LVDS0_CLK  IMX6QDL_CLK_PLL2_PFD0_352M
 #define CONFIG_EFUSA9_LVDS0_CLK  IMX6QDL_CLK_PLL5_VIDEO_DIV
 #define CONFIG_EFUSA9_LVDS1_CLK  IMX6QDL_CLK_PLL2_PFD0_352M
 //#define CONFIG_EFUSA9_LVDS1_CLK  IMX6QDL_CLK_PLL5_VIDEO_DIV

Because we are modifying the assigned-clock settings with this commit
anyway, also add a config macro for audio clock speed. On some
Solo/DualLite boards, the audio clock, which is 24.576 MHz by default,
can be reduced to 12.288 MHz to reduce EMC. Unfortunately this is not
possible on Dual/Quad, because an additional divider is missing there.