Add support for UART, ETH, LEDs and audio to armStoneA5
The following changes take place in the armstonea5.dts device tree file.

- Add uart-has-rtscts for uart1
- Add clock settings for CKO1 + RMII_CLKIN or RMII_CLKOUT depending on
  board revision
- Add node leds for Status1 and Status2 LEDs
- Enable audio, define audio clock for CKO2, move sai2grp next to the
  i2c5_gpio_grp so that they can both be guarded by the same #ifdef

FIXME:
lpuart has problems running on Vybrid when transmit is done with DMA. Maybe
this is not possible if the console is running on the same UART. So disable
TX DMA on Vybrid for now. We have to check this again some day.