Improve imx-uart driver
The imx uart driver automatically enables dma for all non-console uarts.
This adds an 4kB buffer, which delays the transmission of the received
data to the userspace. Also, parity detection is not implemented for
dma-uart.
By adding "no-dma" to a uart entry in the devicetree, the uart will now
be interrupt-driven, will not buffer the data and will support parity
detection