Handle several fb references in PWM backlight driver
The PWM backlight driver can have an optional property "fb-names". This
property holds a list of strings referring to the names of those
framebuffers that are associated with this backlight. This is useful
if a system has more than one display. If no such property is given,
the backlight will respond to all FB_BLANK notifications, no whether
from what disaply they came. But if the backlight should only respond to
a specific display, then simply give the name of the according
framebuffer in fb-names and then the backlight will only respond to
FB_BLANK notificiations of this specific display and ignores all
others.

- Remove function pwm_backlight_check_fb_dt() and move the check
  for fb-names to function pwm_backlight_check_fb_name() that
  did the same for a single name.
- Move parsing of fb-names from pwm_backlight_parse_dt() to
  pwm_backlight_probe() to avoid needing a similar entry in
  struct platform_pwm_backlight_data. There was similar code
  anyway.