Use meaningful framebuffer names im mxc_ipuv3_fb.c
In the past, the framebuffer names given by the mxc_ipuv3_fb.c driver
were rather strange:

IPU 1, DI 0: DISP3_BG
IPU 1, DI 1: DISP3_BG - DI1
IPU 2, DI 0: DISP4_BG
IPU 2, DI 1: DISP4_BG - DI1

An overlay window used "FG" instead of "BG". To be able to use the
framebuffer name more easily as reference, for example in the PWM
backlight for the fsb-names property, we need more specific names
that do not depend on the frambuffer number, but on the function.

So now we use the following concept:

- The framebuffer for the LCD port takes the name "lcd".
- The framebuffer for the LVDS channel 0 takes the name "ldb0".
- The framebuffer for the LVDS channel 1 takes the name "ldb1".
- The framebuffer for the HDMI port takes the name "hdmi".

Actually this is the ID given in property "disp_dev" of the according
mxcfb node in the device tree, extended by the display number in case
of ldb. As LVDS channel 0 can only be used on DI 0 and LVDS channel 1
only on DI 1, the display number is sufficient here (no matter which IPU).

An overlay window gets "_ov" appended to the name.

This results in the following changes:

- To have the needed information when building the framebuffer name, add
  a pointer to the background framebuffer information in function
  mxcfb_setup_overlay().
- In function mxcfb_register(), build the new framebuffer name and store
  it in fbi->fix.id.
- In mxcfb_pan_display(), detect the corresponding alpha channel by just
  looking at IPU and display number, not by looking at the framebuffer
  name.
- Change device trees for all fsimx6 boards to reflect the new names:
  armstonea9qdl.dtsi, armstonea9r2qdl.dtsi, efusa9qdl.dtsi, netdcua9qdl.dtsi
  picomoda9qdl.dtsi, qblissa9qdl.dtsi, qblissa9r2qdl.dtsi.

For now we keep the old variants for reference, disabled by #if 0.