Handle initial state in pwm_bl.c more consistently
The PWM backlight driver assumes that the backlight should be enabled by
default. Only if the backlight is referenced by another node (e.g. by a
display driver), it checks if the corresponding GPIO or regulator is
enabled and decides on this state.

In our point of view, this should be independent from the fact whether
the backlight is referenced by another node or not. If the GPIO to the
backlight power is enabled, or the regulator for the backlight power is
enabled, then the backlight should start enabled, otherwise it should
start disabled.

In addition, the probe function should not set props.power if the
backlight should start disabled. Because then the backlight does not
start by a BL_BLANK_UNBLANK notification alone, it also needs additional
activation of the power entry. This is not what is intended here. Of
course the backlight should start with the UNBLANK notification alone.
So set props.fb_blank instead of props.power, which results in the
intended behaviour.