Do not switch ldb regulator too often on i.MX6 platforms
The ldb (LVDS) driver for i.MX CPUs drivers/video/fbdev/mxc/ldb.c has an
enable and a disable function. These functions are called rather often and
every time the regulator for the ldb power is switched on and off. If
correctly implemented, this would also mean that the backlight needed to
be switched every time. This is nonsense. It is sufficient to switch the
regulator on once when the driver is loaded, i.e. in ldb_init(), and to
switch it off once when te driver is unloaded in ldb_deinit(). This is
how it is also done in the mxc_lcdif.c driver.

- Move the enabling of the regulator from ldb_enable() to ldb_init().
- Move the disabling of the regulator from ldb_disable() to the new
  function ldb_deinit(). Add ldb_deinit() to the .deinit entry of
  ldb_drv, so that it is called by the i.MX display infrastructure.