Swap eth0 and eth1 on AGATEWAY
AGATEWAY only has one Ethernet port, that is located on FEC1, not on
FEC0 as you would expect. Swap the ethernet aliases, so that it will
be shown as eth0 instead of eth1. This also makes sure that the MAC
address given in ethaddr in U-Boot is actually used on this port.