Improve pwm-imx driver - support both polarities correctly
By default the PWM unit will be completely disabled if we use duty_cycle
zero. Then the PWM pin is low. This is the wrong value if we have an
inverted PWM, therefor we need high.
To support both polarities the PWM unit must be enabled the whole time.
So the bit MX3_PWMCR_EN must always be set. We add an additional flag
called "keep-power" which can be set in device-tree to support both
polarities and never shut down the PWM unit.