Add CONFIG_KEEP_VLCD_ON for picocoremx6sx.dts
For the display adapters on PicoCoreBBRGB VLCD is used to power the pca
and touch controller. If defined as regulator lcd-supply, the driver
will power down VLCD, when the display output is changing. This can
conflict with pca and touch. If CONFIG_KEEP_VLCD_ON is defined, VLCD
will never power down.