Improve support for CAN-FD chip MCP251xfd
We already integrated a driver for chip MCP251xfd. This driver was not a
mainline driver. In the meantime in mainline kernel 5.10 there is a
driver available to support chip MCP251xfd. This driver is more compact
then the previous one. So backport mainline driver mcp251xfd to our
kernel. Due to this backport some functions in the can infrastructure
must be implemented to support this driver. Remove the previous
MCP251xfd driver.