Improve display backlight support
The backlight drivers flickers for a moment when it goes from off to on
(value != 0). This is caused by several problems. One problem is the
pwm_bl.c driver which sets the state to disable. This causes that the
pwm control register gets cleared and the clock disabled. The values,
period and duty_cycle are the same before backlight off. So if backlight
gets enabled the display takes the brightness from duty_cycle and period
value before. Thats why we first set the duty_cycle value to logic zero
and display the clock.
A second problem is that the driver pwm-imx27.c set all register value
to zero if backlight gets off. This causes also a small flicker before
the reset bit is set of this register (MX3_PWMCR). Due to this issue
reason we keep the register values of MX3_PWMCR if backlight  gets
disabled.