Improve ksz9477 switch driver to support phy-link via GPIOs
Add possibility to setup LED pins of the switch as GPIOs. The GPIOs can
be configured to act like a link trigger. If link is up GPIO set low, if
link is down GPIO set high. It is also possible to setup either LEDX_0
or LEDX_1, to act like a link trigger.

Move device-tree parsing to specific ksz9477.c instead of ksz_common.c.