Add driver for lontium-lt9211 MIPI2RGB converter
This driver is based on the example driver by Lontium
LT9211_MIPI2TTL_LVDS_Code_v1_4_2-modify20201205.

Because the lt9211 needs valid MIPI input data to configure itself, we
initialize the MIPI side with the simple panel driver and wait in the
lt9211-probe until the panel is enabled.

We do not have implemented any suspend/resume functions yet so this
driver will not work with deep sleep applications.