Fix imx6 device tree pathes for backwards compatibility
Add the leading zeros of some device tree paths again for backwards
compatibility with uboots < fsimx6-B2021.10.
Older Uboots use the absolute path for making adaptions at runtime to
the Linux device tree, which leads to incompatibility, when the Linux
device tree paths are changed (like removing leading zeros).
In future releases we will use aliases instead of absolute paths to
avoid this issue.

This commit may be reversed as soon as the backwards compatibility is
not needed anymore.