Add fsimx6sx/ul alias for gpmi/emmc/rpmsg
UBoot references the gpmi/emmc/rpmsg node by an absolute path. This may
lead to incompatibility between UBoot and Linux if the Linux device tree
path changes.

This can be avoided by referencing aliases in UBoot, which are less
likely to change and can be handled in F&S device trees.