Fix armv7 default cpu frequency governor
Previously, the CPU Freq governor was set to "interactive" by default.
This is a power saving mode, which will only use high frequencies on
high loads and user input.

This mode has been removed in kernel 5.4.3.2.0.0, so now the default
value "performance" is used, which will alway use the maximal frequency
available.
This leads to a higher energy consumption.

Now we use the "ondemand" governor, because it should be the closest to
the old "interactive" governor.