Handle USB Type-C ALERT on J2_77 on PicoCoreMX8MM/MN
The USB Type-C ALERT signal is basically available on J3_15 of the extra
30-pin USB-3 connector on some modules like PicoCoreMX8MP. However if Type-C
is to be used with USB-2.0 on boards without J3, like the PicoCoreMX8MM or
PicoCoreMX8MN, this pin is missing. Starting with Board-Rev 1.40 of the
PicoCoreBBDSI baseboard, this signal is also available on the alternative
pin J2_77.

Handle this pin correctly on PicoCoreMX8MM and PicoCoreMX8MN:

- On PicoCoreMX8MM-DDR3L, this is handled by CPU pad SAI1_RXFS.
- On PicoCoreMX8MM-LPDDR4, this is handled by CPU pad SAI1_RDX6.
- On PicoCoreMX8MN-DDR3L, this pin is not connected and cannot be handled.
- On PicoCoreMX8MN-LPDDR4, this pin is handled by SAI2_RXFS, but only if
  CAN is not equipped. If CAN is equipped, J2_77 is not connected and
  ALERT cannot be handled.

In all cases when ALERT is not available, do not define the interrupt
property in device tree node tcpc@52. Then the driver will automatically
use polling with an interval of ~100ms. Please note that this will
generate quite some traffic on the I2C bus.

Also enable the TCPCI driver (nxp,ptn5110) in fsimx8mm_defconfig.

Also fix some (commented) undefs in picocoremx8mm-ddr3l.dts that still
showed CONFIG_PICOCOREMX8MM_* instead of CONFIG_PICOCOREMX8MX_*.