Improve fbdev mipi_dsi_northwest driver
The EE0350ET MIPI display did not recive the SLEEP_OUT command sometimes
and stayed black ager a blank.
When setting the MIPI clock to LP_MODE explicitly when going to sleep in
mode, the SLEEP_OUT command is received reliably.

Also reduce the mdelay in the mipi_dsi_pkt_write function to 1ms or else
a resume from deep sleep with display takes up to 3 seconds.

It seems to work with 1ms so we will go with that for now.