Use default timings for J070WVTC0211
For the J070WVTC0211 display we generated the timings probably first on
the MIPI2LVDS adapter, which requires a higher pixelclock because of big
steps in the resolution for the videoclock. For boards with a direct
LVDS output, we can and should use the default timings for the display
to achieve a frame rate of 60Hz (previous frame rate was 76Hz).