Fix sec_mipi_dsim initialization with lvds bridge
The imx sec-dsim driver throws an error if the MIPI_DSI_MODE_VIDEO_BURST
or the MIPI_DSI_MODE_VIDEO_SYNC_PULSE flag is not set. We only use the
burst mode so far so we set it in the tc358775 driver

Also the sec_mipi_dsim driver does not clean up correctly when a
EPROBE_DEFER error occurs, and throws errors when trying to probe the
second time.
Add the necessary clean up functions.