Improve picocoremx8mm device tree layout
Because of the increasing amount of board versions with very similar
code base, we need to make as much code as generic as possible.

For this, we create an a generic picocoremx8mm.dtsi, from which other
version like the picocoremx8mm, picocoremx8mmr2 or the picocoremx8mx can
be derived.

Also, we try to make the display device trees as generic as possible, so
the board specific display device trees only have to set some defines
(like gpio number ot i2c bus) and include the display.dtsi and their
board specific device tree.

Also we remove some display device trees that are not supported anymore.