Improve fsimx8mp sound drivers
Use the FSL_ASOC_CARD driver on all fsimx8mp platforms.
This is necessary because the old simple-audio-card driver is not
supported anymore.

Add init-on-array functionality the the clk-blb-ctrl driver to activate
the MCLK1 clock early so the SGTL500 i2c can be used.

It is necessary to set the correct clock parent in this stage, because
the driver cannot reassigned later.