Add GAL1 DTS
This commit adds Devicetree and Makefiles for GAL1 board.
Currently this DT file is used as a template and all entries are not tested on board.
So there will be some changes and adjustments in the future

ATM some GPIOs are declared as HOG's. This is because it is currently not clear how these signals would have to be stored in the device node.