thermal:imx8mm:Improve imx8mm_thermal driver
thermal:imx8mm:Improve functionality of imx8mm_thermal driver to
read temperature.

NXP uses TMU driver in UBoot to enable some thermal functionality
for the kernel, rootfs. Without the driver is it not possible to
read the temperature of the CPU or SOC in rootfs.
The TMU driver in UBoot is used in the early board_f phase by
by calling print_cpuinfo(). In this boot phase no global variables
should be used. However probing the TMU driver violates this rule and
causes damages to the device tree. So do not use CONFIG_NXP_TMU for now.

Backported from nxp branch lf-5.15.y tag lf-5.15.32-2.0.0