Improve bdinfo driver
We want to be able to read the board-cfg and board-id from linux.

For this the bdinfo driver  needs to be able to also parse subnodes for
the bdinfo device tree node.

The Uboot needs to add the board-cfg to the bdinfo device tree node as a
subnode.
Then the bdinfo driver can parse the subnode and add it as a
subdirectory to the sysfs add /sys/bdinfo.

This will currently work only with one subnode layer, so if we add an
additional subnodes to the board-cfg, we will have to adapt the bdinfo
driver again.

A simple heuristic will try to determine the data type of each device
tree entry. This may fail for certain circumstances, but should be ok
most of the time.
The heuristic is taken from
scripts/dtc/fdtget.c
and
scripts/dtc/fdtdump.c