Add F&S Linux Version
We want to add the git version of the Linux Kernel to the binary.

Normally this would be done with setlocaversion and the UTS_RELEASE
string.

But this will break the loading of modules if only the kernel is
exchanged and the rootfs stays the same, because the UTS_RELEASE string
is taken to match the modules-directory name with the Kernel Release
name.

So we introduce our own version string that will be printed at kernel
boot and can be read out via bdinfo.