Add device tree version
We want to add a device tree version, fsversion.hthat can be read out at runtime.
Add a dts-version entry to the bdinfo node, that will be set to the
current Kernel build version by default.
The user can set dts-version by hand to a own version, too.

The device trees will have to include fsversion.h

and add the dts entry to the bdinfo node:
 	bdinfo {
		dts_version = FS_LINUX_VERSION;
 	};