Fix armstonemx8mp recording issue
The wm8960 is not able to record, because in codec-slave mode we need to
set the ADCLRC pin to GPIO1 to be able to use DACLRC as clock for both
input and output.

We are currently not able to use the simple-audio-card driver in
codec-master mode and for the next kernel version, we are going to
switch to the fsl-asoc-card driver.