IC TPS65185: Add Driver for TI EPD PMIC device
Author: Claudio Senatore <senatore@fs-net.de>  2023-03-03 13:52:27

Implement drivers for Texas Instrument's EPD PMIC family TPS65185.

Three drivers are provided for the Linux kernel that are intended
for
    the PMIC.

    MFD driver:
The MFD driver handles the I2C interface of the device. Here
functions
    are defined, which read and write the registers. When the driver is
    loaded, the operating state of the PMIC is changed from SLEEP to
STANDBY. When the driver is unloaded, the operating state is reset
to
    SLEEP.
In the probe function, two more subdevices are registered for the
PMIC,
    which are handled by the tps65185-pmic and tps65185-hwmon drivers.

    Regulator driver:
    The tps65185-regulator driver handels tps65185-pmic subdevice.
It reads the regulators defined in Device-Tree and registers it
within
    probe function. Here are also defined functions to enable, disable,
    set/get voltages for specific regulators that are used by a customer
(EPDC-Driver). The Driver uses GPIO-Powerup and GPIO-Powerstat
signals
    to control the PMIC-statemachine.

    HWMON driver:
The tps65185-hwmon driver creats a sysfs attribute temp1_input
whitch
    reads the temperature from the PMIC Register.

    TO DO:
    The Driver does not handle interrupt signals from PMIC