Add support to disable DTS pinctl nodes
We want to be able to disable pincontrol nodes in the devicetree.
This way we could select different pincontrols dynamically from uboot,
by enabling/disabling the node.

This should not affect any existing device trees as the logic to returtn
True if the status property is absent or set to "okay" or "ok" and false
otherwise.