fs-gal1.dtb: Set GPIOs for mPCIe & M.2
It is currently possible to detect devices connected to mPCIe
(right-handed). This requires some GPIOs to be set

- GPIO "PCIe1_PWR_EN" must be high to enable the mPCIe (right-handed)
PWR-	Switch.
- GPIO "PCIe1_PWR_EN" must be high to enable mPCIe/M.2 (left-handed)
PWR-	Switch.

This is solved with a fixed-regulator with regulator-always-on property.

The Layerscape-PCIe driver does not handle supply (regulator) properties.
To clear a reset of a connected PCIe device, gpio-hogs are used to raise
the PERST signal.

The Layerscape-PCIe driver does not handle clock properties, so therefore the
GPIO "PCIe_CLKGEN" is always pulled down to enable the VERSACLOCK
outputs for mPCIe and M.2.

The sim-switch signal is also pulled down with a gpio-hog to enable
right- and left-handed sim sockets as default.

All GPIO signals that are defined as a gpio-hog or regulator are marked
as "used" and cannot be modified at runtime.

With these changes, PCIe devices can be recognised and the corresponding
drivers automatically loaded.