Prevent writing to UARTMODIR in lpuart32_setup_watermark()
With every open() syscall on a lpuart device, the lpuart32_hw_setup() is
called. The lpuart32_hw_setup() calls lpuart32_setup_watermark_enable()
witch calls lpuart32_setup_watermark().

The lpuart32_setup_watermarks() has a broken section:

	/* set RTS watermark */
	if (!uart_console(&sport->port)) {
		val = lpuart32_read(&sport->port, UARTMODIR);
		val = (sport->rxfifo_size >> 1) << UARTMODIR_RTSWATER_S;
		lpuart32_write(&sport->port, val, UARTMODIR);
	}

Here the rxfifo_size is written in a wrong register and overwrites
TXRTSPOL and TXRTSE. These bits are nessesary for RS485 mode!
Those bits that contains informations about rxfifo_size are written in a
reserved bitfield in UARTMODIR.

In the newest fsl-lpuart.c driver in Torvalds Linux-Git is acctually a
fix with a the Bitfield definition, that prevent the overwrite of
TXRTSPOL and TXRTSE. This solutuion is adapted to current fsl-lpuart.c
driver

For LPUART3 interface, the GPIO signals are defined in fs-gal1.dtb