Fix Realtek Ethernet Phy Bug in Low Power Mode
When starting both Realtek Phys on an r2 model and only one is connected
with an ethernet cable, the connection on the other Phy with the
ethernet cable breaks. If the unconnected Phy is powerd down or gets an
ethernet connection the connection works again.

he cause is an error in the power saving mode. An option is enabled that
causes the second phy to disable a clock, used by the first phy.

Disable it to restore functionality to both the ethernet phys itself and
the power saving mode.