Disable busfreq scaling for picocoremx8mm-ddr3l
There seem to be issues with the bus frequency scaling for picocoremx8mm
with DDR3L ram. The board freezes if display and ethernet are turned
off.

As we do not plan to use ddr3l in the future, we disable bus frequency
scaling for now for picocoremx8mm-ddr3l boards.