Add Secmon Status register for verification into sysfs
To check if the OTPMK fuses are set in a correct way the user needs to
consult the secmon status register. To make it easily accessible add an
attribute file to the sysfs.

Currently the file is located at
/sys/devices/platform/soc/1e80000.efuse/secmon_hp_status. While it would
be favourable to have it at
/sys/bus/nvmem/devices/fsl-sfp0/secmon_hp_status, this would need
changes to the core driver to hand the kobject into the layerscape
driver.