fs-gal1.dts: disable temp sensor
The temp sensor is only used on development boards or on boards with
special configuration definitions. Therefore it will be disabled as
default.

If a board definition contains the sensor, the note is activated via
U-Boot.

The interrupt property is also removed, since the device does not need
to detect any defined temperature thresholds