Set netdev name according to DT label property
The requirement has come up that labels for network interfaces should be
defined via the devicetree. This is already implemented in linux for DSA
devices via the Linux net framework. Unfortunately, the net/ethernet
framework for ethernet devices does not support label declarations for
netdev names. There are also discussions about this in mailing lists
which are mentioned here:

https://lore.kernel.org/netdev/140c3841-51fc-6874-b497-4ead9a623d34@seco.com/T/

or here:

https://patchwork.kernel.org/project/linux-arm-kernel/patch/1399390594-1409-1-git-send-email-boris.brezillon@free-electrons.com/

Nevertheless, a name can be redefined via an Ethernet driver for its
interface. This is an example implementation of how this could look for
freescales enetc drivers.

In the devicetree, the label property can be read out for each port
node. It is possible to use wildcards (e.g "eth%d") to automate the
numbering.
However, it is also possible to define fixed names.

Please note that the interface cannot be registered if the name already
exists for another interface!