Support for OPM103E2 Epaper-Display
Added struct named as opm103E2mode

Added opm103E2mode

Changed is the VCOM to 800000, which is in the datasheet -0.80 V

TODO: There is no specific information in its datasheet about controller
timings. Thus, all of the other parameters have yet to be found for this
Display except xres, yres and vcom.