Fix dp83848 ethernet driver
The dp83848 needed some adaptions to work with kernel 5.15.

Also the phy_write function in dp83848_config_init did not really write
the value, set with the device tree entries.

Also rename the device tree flag dp83848,led-activity to
dp83848,no-led-activity, because by default the activity led should be
on and setting the flag should disable it.