net: phy: realtek: ADD Fiber support for RTL8211FSI
The RTL8211FSI supports various media, such as UTP, fiber and SGMII.

To set up a 1000BASEX Phylink mode via the Serdes interface, adaptations
are required for the probe and the speed announcement.

The SFP driver is included in the probe, which checks the EEPROM of the
SFP module and configures the link modes and speed.

As the register value for page 0xa43 reg 0x12 no longer represents the
speed, a new read_status() function is required.