thermal: qoriq_thermal: register tmu zone for all id's
Since passive cooling is provided for a thermal zone via the device
tree, core frequencies must be reduced to keep the processor cool.

The Thermal Governor Step Wise uses the current temperature, the
temperature trend, and the trip temp as limits to determine whether a
cdev should be clocked up or down.

The LS1028A has two thermal zones for DDR and CPU.
However, the driver only registers the DDR thermal zone, which means
that the trip temp for the CPU is never set and represents the number 0.
As a result, the trent will always rise.

This bug ensures that the processor can clock down at a high
temperature, but is not allowed to clock up in a cooler environment.

The provided change gives a correct result for falling and rising
temperatures for Thermal-Zone1 (CPU).