Fix fsl_ssi pm_runtime warning
The fsl_ssi driver uses pm_runtime_enable() in the probe, but no
pm_runtime_disable().

Because we load the sdma driver as a module, the probe of
fsl_ssi failes with -EAGAIN multiple times until the rootfs and we get a
lot of "Unbalanced pl_runtime_enable!" warnings.

Run pm_runtime_disable() in case of error.