mac80211 (ath10k): Do not advertise CONTROL_PORT_OVER_NL80211
New wpa_supplicants since 2.10 use a control port instead of a socket to do
WPA 4WAY_HANDSHAKE to avoid a rare race condition during authentication.
For some unknown reason, this does not work on the ath10k wireless driver
for Silex. Maybe the driver does not support this correctly. This causes
the 4WAY_HANDSHAKE in the second step to always fall back to the first step
instead of continuing with the third step. So authentication fails with a
timeout after 10 seconds and the station drops the connection to the access
point it is currently trying to associate with.

However wpa_supplicant only does this if the driver announces the feature

  NL80211_EXT_FEATURE_CONTROL_PORT_OVER_NL80211

Remove the flag (and two associated flags) from net/mac80211/main.c. Then
wpa_supplicant will use the old socket method, with the rather small risk
of triggering the race condition. As all other drivers (old Silex driver
from kernel 4.9.88 and mxm-wifiex driver for Azurewave) also do not
announce this feature, this seems to be acceptable.

REMARK:

To check if this is happening, prepare a wpa_supplicant.conf with a
working WLAN network. Then call wpa_supplicant with the following
command line:

  wpa_supplicant -B -D nl80211 -c /etc/wpa_supplicant.conf -i wlan0 -dd \
     -f /tmp/wpa_log.txt

Then wait until an association attempt fails (or succeeds) and then check
the wpa_log.txt file.

Search for

  wlan0: State: ASSOCIATED -> 4WAY_HANDSHAKE

and

  wlan0: WPA: RX message 1 of 4-Way Handshake from ... (ver=2)

If there are the following messages shortly after that, the control port method
is used.

  nl80211: Send over control port dest=... proto=0x888e len=121 no_encrypt=1
  nl80211: tx_control_port cookie=...

If the next step goes back to the following message, authentication fails:

  wlan0: WPA: RX message 1 of 4-Way Handshake from ... (ver=2)

However if the next step shows the following message, authentication succeeds:

  wlan0: WPA: RX message 3 of 4-Way Handshake from ... (ver=2)

FIXME:

Check if this will work on newer kernel versions or if there is a fix for the
ath10k driver. Then this workaround can be removed again.