Improve leds-pca963x-fus driver
The pca963x chips offer support for a lower PWM frequency of 190MHz when
using the group pwm.
Add support to use this frequency by setting the respective period in
the pwm device tree entry.

Also add support for an optional regulator, because some display
adapters supply the pca963x with the the vlcd regulator.