Fix fsimx6 audio
Activate SND_SOC_FSL_SSI  in the defconfig. Also don't select SND_SOC_FSL_SSI or
SND_SOC_FSL_SAI with SND_SOC_IMX_SGTL5000 anymore. Move it to the
defconfigs instead

Also, make sure that the ssi driver is loaded before loading the codec or else we run into
an infinite probe-defer-loop while waiting for the imx_sdma driver, which is build as a module