Fix RPMSG failure on boot
When booting with a running M-Core and RPMSG activeted in the Linux
Devicetree, booting fails if there is no rsc_table initialized. The
remoteproc driver does not check if there is a table but assumes it is
there and uses wrong values and tries to run outside of existing memory.

Fix this behaviour by checking the table version field and acting
accordingly.