Fix imx uart dma watermark level
It was not possible to read data from serial port when input buffer
contains exactly 9 bytes.
The number 9 is related to the length of the rx_fifo_trig in
drivers/tty/serial/imx.c. The src_maxburst has to be one byte smaller
than the watermark level to trigger the aging timer.

This fix is available since kernel 4.4 but did not get merged properly

We restrict the "fsl,rx_fifo_trig" to a minimum of Bytes, so the
src_maxburst can still be one byte smaller. If it has to be smaller
"no-dma" has to be used.