drivers:rtc:rtc-pcf85363: add clk-out and drive property
The pcf85263 and pcf85363 provide registers for configuring drive
strength and output clk. Depending on the hardware design,
adjustments may be required.

This commit complements the "clk-out" and "quartz-drive-strength"
properties from the device tree. This allows board-specific
RTC-adjustments to be made via the device tree without having to
adjust the driver individually.