fsimx6: Fix ipu_pixel_clk parent recognition
When the display has already been activated in U-Boot, the mux for the
ipu_di_pixel clock (in DI_GENERAL of the IPU registers) may be set to 1,
i.e. the external display clock ipu1_di. However the code in
drivers/mxc/ipu3/ipu_pixel_clk.c does not detect this. It silently
assumes an initial setting of 0 and initializes mux->index accordingly.
Therefore the first call to _ipu_pixel_clk_get_parent(), when registering
the clk object, will always return the internal clock ipu1, even if the
clock is actually set to the external clock ipu1_di.

Later in ipu_init_sync_panel() (in drivers/mxc/ipu3/ipu_disp.c), the clock
parent is set to the internal ipu1 clock. However the clk infrastructure
sees no change in the parent and does not do anything at all, i.e. it does
not call _ipu_pixel_clk_set_parent(). This keeps the wrong parent active,
i.e. ipu1_di instead of ipu1.

At the end of the boot process, when inactive clocks are disabled, the
ipu1_di clock apparently has no user and is disabled, resulting in the
real display clock being switched off. Even the clock tree shows the wrong
hierarchy here (in /sys/kernel/debug/clk/clk_summary).

To fix this, read the current register setting in clk_register_mux_pix_clk()
and initialize mux->index to 1 if the external clock ipu1_di is currently
active. Here, we can savely assume that the IPU is online and the register
can be read. Doing this in _ipu_pixel_clk_get_parent() would be no good idea
as this function may also be called when the IPU is offline. This is why
there only mux->index is returned instead of actually reading the register.